+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Added `VueObserveVisibility` and `VueLazyload` dependency
+- Added safe Area of Elements in the Bottom Elements(#499)
 
 ### Changed / Improved